Refactor: forward-declare RequestManagerRequest
