+++ src/groups/bmq/bmqp/bmqp_requestmanager.h
@@ -19,8 +19,10 @@
 //@PURPOSE: Provide a mechanism to manipulate requests and their response.
 //
 //@CLASSES:
-//  bmqp::RequestManagerRequest: request and its associated context
-//  bmqp::RequestManager:        mechanism to manage requests and responses
+//  bmqp::RequestManager: mechanism to manage requests and responses
+//
+//@SEE_ALSO:
+//  bmqp_requestmanagerrequest: request and its associated context
 //
 //@DESCRIPTION: 'bmqp::RequestManager' is a mechanism to manage requests (using
 // the 'bmqp::RequestManagerRequest' type) and their associated responses.
@@ -67,12 +69,11 @@
 //
 /// Thread Safety
 ///-------------
-// The 'bmqp::RequestManager' and 'bmqp::RequestManagerRequest' classes are
-// fully thread-safe (see 'bsldoc_glossary'), meaning that two threads can
-// safely call any methods on the *same* *instance* without external
-// synchronization, and re-entrant safe, meaning that a method of this object
-// can be called from within a callback emanating from that same object
-// instance.
+// The 'bmqp::RequestManager' class is fully thread-safe (see
+// 'bsldoc_glossary'), meaning that two threads can safely call any methods on
+// the *same* *instance* without external synchronization, and re-entrant safe,
+// meaning that a method of this object can be called from within a callback
+// emanating from that same object instance.
 //
 /// Request Cancellation Order
 ///--------------------------
@@ -95,20 +96,11 @@
 //
 /// Distributed Trace Integration
 ///-----------------------------
-// An externally created 'bmqpi::DTSpan' may be attached to a request using the
-// 'RequestManagerRequest::setDTSpan' method. This ensures that a span
-// representative of a request survives for at least the lifetime of the
-// request object.
-//
-// A 'RequestManagerRequest' which owns a 'DTSpan' can additionally be provided
-// with a 'bmqpi::DTContext' via the 'RequestManagerRequest::setDTContext'
-// method. This guarantees that the 'DTSpan' will be the active span of the
-// 'DTContext' whenever a response- or signal- callback of the request is
-// invoked. The easiest (and recommended) way to attach a 'DTContext' is by
-// using the 'RequestManager::setDTContext' method on the object used to create
-// new 'RequestManagerRequest' instances; the context will be propagated to any
-// new requests created. Note that setting a 'DTContext' without a 'DTSpan'
-// will have no effect.
+// A 'bmqpi::DTContext' provided at construction is propagated to every
+// 'bmqp::RequestManagerRequest' created by this object, so that a span
+// attached to a request is activated within that context whenever a callback
+// for the request is invoked.  See the 'bmqp_requestmanagerrequest' component
+// for details.
 //
 /// Usage Example (basic)
 ///---------------------
@@ -325,319 +317,51 @@
 
 // BMQ
 
-#include <bmqp_ctrlmsg_messages.h>
+#include <bmqp_blobpoolutil.h>
 #include <bmqp_protocol.h>
 #include <bmqp_schemaeventbuilder.h>
-#include <bmqpi_dtcontext.h>
-#include <bmqpi_dtspan.h>
 #include <bmqt_resultcode.h>
 
 #include <bmqc_orderedhashmap.h>
 #include <bmqex_executor.h>
-#include <bmqex_systemexecutor.h>
-#include <bmqio_channel.h>
-#include <bmqio_status.h>
-#include <bmqu_memoutstream.h>
-#include <bmqu_printutil.h>
-#include <bmqu_time.h>
 
 // BDE
 #include <ball_log.h>
-#include <bdlbb_blob.h>
-#include <bdld_manageddatum.h>
-#include <bdlf_bind.h>
-#include <bdlf_placeholder.h>
-#include <bdlmt_eventscheduler.h>
-#include <bsl_cstddef.h>
 #include <bsl_functional.h>
-#include <bsl_iostream.h>
 #include <bsl_limits.h>
 #include <bsl_memory.h>
 #include <bsl_string.h>
-#include <bsl_utility.h>
-#include <bsla_annotations.h>
 #include <bslma_allocator.h>
-#include <bslma_managedptr.h>
 #include <bslma_usesbslmaallocator.h>
-#include <bslmf_allocatorargt.h>
 #include <bslmf_nestedtraitdeclaration.h>
-#include <bslmt_lockguard.h>
 #include <bslmt_mutex.h>
-#include <bslmt_semaphore.h>
-#include <bsls_assert.h>
 #include <bsls_cpp11.h>
-#include <bsls_performancehint.h>
-#include <bsls_systemclocktype.h>
 #include <bsls_timeinterval.h>
-#include <bsls_timeutil.h>
 #include <bsls_types.h>
 
 namespace BloombergLP {
 
+// FORWARD DECLARATION
+namespace bdlbb {
+class Blob;
+}
+namespace bdlmt {
+class EventScheduler;
+}
+namespace bmqio {
+class Channel;
+}
+namespace bmqp_ctrlmsg {
+class ControlMessage;
+}
+namespace bmqpi {
+class DTContext;
+}
+
 namespace bmqp {
 
 // FORWARD DECLARATION
-class RequestManager;
-
-// ================================
-// struct RequestManagerComponentId
-// ================================
-
-/// Integer constants used to tag requests sent via a `RequestManager` for the
-/// purpose of component-scoped cancellation.
-struct RequestManagerComponentId {
-    // CONSTANTS
-    enum {
-        /// Default; no component association.
-        k_NO_COMPONENT_ID = 0,
-
-        /// Cluster FSM component id.
-        k_CLUSTER_FSM = 1,
-
-        /// Base value for Partition FSM component ids.  Each partition
-        /// gets a unique componentId computed as
-        /// `k_PARTITION_FSM_PREFIX + partitionId`.  Use the
-        /// `partitionFSM()` class method to obtain the componentId for a
-        /// given partition.
-        k_PARTITION_FSM_PREFIX = 900
-    };
-
-    // CLASS METHODS
-
-    /// Return the componentId for the partition with the specified
-    /// `partitionId`.  The behavior is undefined unless
-    /// `0 <= partitionId < 100`.
-    static int partitionFSM(int partitionId);
-
-    /// Return true if the specified `componentId` is valid, including
-    /// `k_NO_COMPONENT_ID`.
-    static bool isValid(int componentId);
-};
-
-// ===========================
-// class RequestManagerRequest
-// ===========================
-
-/// Object representing a request sent, pending response for it; holding the
-/// request and all associated context state.  This allows for both
-/// synchronous and asynchronous response management.
-class RequestManagerRequest {
-  public:
-    // PUBLIC CLASS DATA
-    static const int k_NO_GROUP_ID = -1;
-
-    // TYPES
-
-    /// Signature of a callback for delivering the response in the
-    /// specified `context`.
-    typedef bsl::function<void(
-        const bsl::shared_ptr<RequestManagerRequest>& context)>
-        ResponseCb;
-
-    /// Signature of a callback for signaling a response in the specified
-    /// `context`.
-    typedef bsl::function<void(
-        const bsl::shared_ptr<RequestManagerRequest>& context)>
-        AsyncNotifierCb;
-
-  private:
-    // DATA
-    bsl::weak_ptr<RequestManagerRequest> d_self_wp;
-    // Weak pointer to self
-
-    bmqp_ctrlmsg::ControlMessage d_requestMessage;
-    // The request
-
-    bmqp_ctrlmsg::ControlMessage d_responseMessage;
-    // The response
-
-    bslmt::Semaphore d_semaphore;
-    // Semaphore associated to this request,
-    // used for synchronous calls wait.
-
-    bdlmt::EventScheduler::EventHandle d_timeoutSchedulerHandle;
-    // Scheduler handle for the timeout
-    // associated to this request.
-
-    ResponseCb d_responseCb;
-    // Response callback, if any, to invoke
-    // upon reception of a response for this
-    // request.
-
-    AsyncNotifierCb d_asyncNotifierCb;
-    // Callback invoked when calling signal,
-    // if it exists.
-
-    bsls::Types::Int64 d_sendTime;
-    // Time when the request was sent, used
-    // for statistics/logging.
-
-    bsl::string d_nodeDescription;
-    // Description of the node the request
-    // was sent to.
-
-    bool d_haveTimeout;
-    // Whether the timeout for the request
-    // has been invoked.
-
-    bool d_haveResponse;
-    // Whether a response for teh request has
-    // been received.
-
-    int d_groupId;
-    // The 'groupId' associated with this
-    // request. It is used when canceling
-    // requests, allowing to cancel all
-    // requests sharing the same group.
-
-    int d_componentId;
-    // The 'componentId' associated with this
-    // request.  Used to cancel all requests
-    // belonging to a specific component.
-
-    bsl::shared_ptr<bmqpi::DTSpan> d_dtSpan_sp;
-    // Distributed Trace span representing
-    // this request.
-
-    bmqpi::DTContext* d_dtContext_p;
-    // A Distributed Trace context which will
-    // take 'd_dtSpan_sp' as the active span
-    // for the duration of any callbacks
-    // executed on behalf of this request. If
-    // either this or 'd_dtSpan_sp' are null,
-    // then 'd_dtSpan_sp' will not be the
-    // active span for any context.
-
-    bdld::ManagedDatum d_userData;
-    // Optional userData associated with this
-    // request.
-
-    // FRIENDS
-    friend class RequestManager;
-
-  private:
-    // NOT IMPLEMENTED
-    RequestManagerRequest(const RequestManagerRequest&) BSLS_CPP11_DELETED;
-
-    /// Copy constructor and assignment operator are not implemented.
-    RequestManagerRequest&
-    operator=(const RequestManagerRequest&) BSLS_CPP11_DELETED;
-
-    /// If a Distributed Trace span and context have been set for this
-    /// request, then this sets the span as the active one within the
-    /// context, returning a `token` that will revert the context's state
-    /// upon its destruction. If either no span has been set for this
-    /// request, or no context has been set for the span to be made
-    /// active within, then an empty `ManagedPtr` is returned.
-    bslma::ManagedPtr<void> activateDTSpan() const;
-
-  public:
-    // TRAITS
-    BSLMF_NESTED_TRAIT_DECLARATION(RequestManagerRequest,
-                                   bslma::UsesBslmaAllocator)
-
-    // CREATORS
-
-    /// Create a new object using the specified `allocator`
-    explicit RequestManagerRequest(bslma::Allocator* allocator);
-
-    /// Destroy this object
-    ~RequestManagerRequest();
-
-    // MANIPULATORS
-
-    /// Clear this object to a default constructed state.
-    void clear();
-
-    /// Wakes up any thread that was waiting on this request (from a call to
-    /// `wait()`) by positing to the semaphore.  This will also invoke the
-    /// `asyncNotifierCb` if one was provided.
-    void signal();
-
-    /// Block until a response for this request is received (or until the
-    /// request times out or gets canceled).
-    void wait();
-
-    bmqp_ctrlmsg::ControlMessage& request();
-
-    /// Return a reference offering modifiable access to the corresponding
-    /// member of this object.
-    bmqp_ctrlmsg::ControlMessage& response();
-
-    RequestManagerRequest& setResponseCb(const ResponseCb& value);
-
-    /// Set the corresponding member to the specified `value` and return a
-    /// reference offering modifiable access to this object.
-    RequestManagerRequest& setAsyncNotifierCb(const AsyncNotifierCb& value);
-
-    /// Set group identifier to the specified `value` to assist canceling
-    /// requests belonging to one group only.  By default, the identifier is
-    /// `k_NO_GROUP_ID` meaning the request does not belong to any group in
-    /// which case it gets cancelled by `cancelAllRequests` without
-    /// `groupId`.  Return a reference offering modifiable access to this
-    /// object.
-    RequestManagerRequest& setGroupId(int value);
-
-    /// Set component identifier to the specified `value` to assist canceling
-    /// requests belonging to one component only.  By default, the identifier
-    /// is `RequestManagerComponentId::k_NO_COMPONENT_ID`.  Return a
-    /// reference offering modifiable access to this object.  The behavior is
-    /// undefined unless `value` is valid and is not `k_NO_COMPONENT_ID`.
-    RequestManagerRequest& setComponentId(int value);
-
-    /// Take shared ownership of the specified `span` and ensure that it
-    /// lives at least as long as this object. The `span` is intended to
-    /// represent this request.
-    ///
-    /// If there is a Distributed Trace context (set via `setDTContext`),
-    /// then `span` will be made its active span for the duration of any
-    /// callbacks executed on behalf of this request.
-    RequestManagerRequest&
-    setDTSpan(const bsl::shared_ptr<bmqpi::DTSpan>& span);
-
-    /// Stores an unowned pointer to `context`: If a span has been set (via
-    /// `setDTSpan`), then it will be made the active span of `ctx` for the
-    /// duration of any callbacks executed on behalf of this request.
-    RequestManagerRequest& setDTContext(bmqpi::DTContext* ctx);
-
-    /// Take ownership of the specified `value` and destroy the user data
-    /// previously managed by this object.  The behavior is undefined unless
-    /// `value` was allocated using the same allocator used by this object
-    /// and is not subsequently destroyed externally using `Datum::destroy`.
-    RequestManagerRequest& adoptUserData(const bdld::Datum& value);
-
-    // ACCESSORS
-    bool                                isLateResponse() const;
-    bool                                isLocalTimeout() const;
-    bool                                isError() const;
-    const bmqp_ctrlmsg::ControlMessage& request() const;
-    const bmqp_ctrlmsg::ControlMessage& response() const;
-
-    /// Return the value of the corresponding member.
-    const ResponseCb& responseCb() const;
-
-    /// Convenient accessor to return the `GenericResult` of the response
-    /// associated to this object: return either `success` or the category
-    /// code associated to the Failure type of the response.
-    bmqt::GenericResult::Enum result() const;
-
-    /// Convenient accessor to return the domain specific error code of the
-    /// response associated to this object: return either `success` or the
-    /// error code associated to the Failure type of the response.
-    int statusCode() const;
-
-    /// Return the description of the node the request was sent to.
-    const bsl::string& nodeDescription() const;
-
-    /// Return the associated group id (`NO_GROUP_ID` by default).
-    int groupId() const;
-
-    /// Return the associated component id (`k_NO_COMPONENT_ID` by default).
-    int componentId() const;
-
-    /// Return the associated user data.
-    const bdld::Datum& userData() const;
-};
+class RequestManagerRequest;
 
 // ====================
 // class RequestManager
@@ -882,745 +606,6 @@ class RequestManager {
     void cancelComponentRequests(const bmqp_ctrlmsg::ControlMessage& reason,
                                  int componentId);
 };
-
-// ============================================================================
-//                             INLINE DEFINITIONS
-// ============================================================================
-
-// --------------------------------
-// struct RequestManagerComponentId
-// --------------------------------
-
-inline int RequestManagerComponentId::partitionFSM(int partitionId)
-{
-    // PRECONDITIONS
-    BSLS_ASSERT_SAFE(partitionId >= 0);
-    BSLS_ASSERT_SAFE(partitionId < 100);
-
-    return k_PARTITION_FSM_PREFIX + partitionId;
-}
-
-inline bool RequestManagerComponentId::isValid(int componentId)
-{
-    return componentId == k_NO_COMPONENT_ID || componentId == k_CLUSTER_FSM ||
-           (componentId >= k_PARTITION_FSM_PREFIX &&
-            componentId < k_PARTITION_FSM_PREFIX + 100);
-}
-
-// ---------------------------
-// class RequestManagerRequest
-// ---------------------------
-
-inline RequestManagerRequest::RequestManagerRequest(
-    bslma::Allocator* allocator)
-: d_requestMessage(allocator)
-, d_responseMessage(allocator)
-, d_semaphore()
-, d_timeoutSchedulerHandle()
-, d_responseCb(bsl::allocator_arg, allocator)
-, d_asyncNotifierCb(bsl::allocator_arg, allocator)
-, d_sendTime(0)
-, d_nodeDescription(allocator)
-, d_haveTimeout(false)
-, d_haveResponse(false)
-, d_groupId(k_NO_GROUP_ID)
-, d_componentId(RequestManagerComponentId::k_NO_COMPONENT_ID)
-, d_dtSpan_sp(NULL)
-, d_dtContext_p(NULL)
-, d_userData(allocator)
-{
-    // NOTHING
-}
-
-inline RequestManagerRequest::~RequestManagerRequest()
-{
-    clear();
-}
-
-inline void RequestManagerRequest::clear()
-{
-    // PRECONDITIONS
-    BSLS_ASSERT_SAFE(!d_timeoutSchedulerHandle);
-
-    // Reset the semaphore
-    while (d_semaphore.tryWait() == 0) {
-        // nothing
-    }
-
-    // d_self_wp: Purposely not reset-ed :)
-    d_requestMessage.reset();
-    d_responseMessage.reset();
-    d_responseCb      = bsl::nullptr_t();
-    d_asyncNotifierCb = bsl::nullptr_t();
-    d_sendTime        = 0;
-    d_nodeDescription.clear();
-    d_groupId     = k_NO_GROUP_ID;
-    d_componentId = RequestManagerComponentId::k_NO_COMPONENT_ID;
-    d_dtSpan_sp.reset();
-    d_dtContext_p = NULL;
-}
-
-inline void RequestManagerRequest::signal()
-{
-    if (d_asyncNotifierCb) {
-        bsl::shared_ptr<RequestManagerRequest> context = d_self_wp.lock();
-        BSLS_ASSERT_SAFE(context);
-
-        bslma::ManagedPtr<void> spanToken(context->activateDTSpan());
-        d_asyncNotifierCb(context);
-    }
-
-    d_semaphore.post();
-}
-
-inline void RequestManagerRequest::wait()
-{
-    // No need to timedWait on a timedSemaphore, 'sendRequest()' schedules an
-    // event for timeout that will post on the semaphore.
-    d_semaphore.wait();
-}
-
-inline bmqp_ctrlmsg::ControlMessage& RequestManagerRequest::request()
-{
-    return d_requestMessage;
-}
-
-inline bmqp_ctrlmsg::ControlMessage& RequestManagerRequest::response()
-{
-    return d_responseMessage;
-}
-
-inline RequestManagerRequest&
-RequestManagerRequest::setResponseCb(const ResponseCb& value)
-{
-    d_responseCb = value;
-    return *this;
-}
-
-inline RequestManagerRequest&
-RequestManagerRequest::setAsyncNotifierCb(const AsyncNotifierCb& value)
-{
-    d_asyncNotifierCb = value;
-    return *this;
-}
-
-inline RequestManagerRequest& RequestManagerRequest::setGroupId(int value)
-{
-    d_groupId = value;
-    return *this;
-}
-
-inline RequestManagerRequest& RequestManagerRequest::setComponentId(int value)
-{
-    // PRECONDITIONS
-    BSLS_ASSERT_SAFE(RequestManagerComponentId::isValid(value));
-    BSLS_ASSERT_SAFE(value != RequestManagerComponentId::k_NO_COMPONENT_ID);
-
-    d_componentId = value;
-    return *this;
-}
-
-inline RequestManagerRequest&
-RequestManagerRequest::setDTSpan(const bsl::shared_ptr<bmqpi::DTSpan>& span)
-{
-    d_dtSpan_sp = span;
-    return *this;
-}
-
-inline RequestManagerRequest&
-RequestManagerRequest::setDTContext(bmqpi::DTContext* ctx)
-{
-    d_dtContext_p = ctx;
-    return *this;
-}
-
-inline RequestManagerRequest&
-RequestManagerRequest::adoptUserData(const bdld::Datum& value)
-{
-    d_userData.adopt(value);
-    return *this;
-}
-
-inline bslma::ManagedPtr<void> RequestManagerRequest::activateDTSpan() const
-{
-    bslma::ManagedPtr<void> result;
-    if (d_dtSpan_sp && d_dtContext_p) {
-        result = d_dtContext_p->scope(d_dtSpan_sp);
-    }
-    return result;
-}
-
-inline bool RequestManagerRequest::isLateResponse() const
-{
-    return d_haveTimeout && d_haveResponse;
-}
-
-inline bool RequestManagerRequest::isLocalTimeout() const
-{
-    return d_haveTimeout && !d_haveResponse;
-}
-
-inline bool RequestManagerRequest::isError() const
-{
-    return isLocalTimeout() ? true : response().choice().isStatusValue();
-}
-
-inline const bmqp_ctrlmsg::ControlMessage&
-RequestManagerRequest::request() const
-{
-    return d_requestMessage;
-}
-
-inline const bmqp_ctrlmsg::ControlMessage&
-RequestManagerRequest::response() const
-{
-    return d_responseMessage;
-}
-
-inline const RequestManagerRequest::ResponseCb&
-RequestManagerRequest::responseCb() const
-{
-    return d_responseCb;
-}
-
-inline bmqt::GenericResult::Enum RequestManagerRequest::result() const
-{
-    if (d_responseMessage.choice().isStatusValue()) {
-        return static_cast<bmqt::GenericResult::Enum>(
-            d_responseMessage.choice().status().category());  // RETURN
-    }
-
-    return bmqt::GenericResult::e_SUCCESS;
-}
-
-inline int RequestManagerRequest::statusCode() const
-{
-    if (d_responseMessage.choice().isStatusValue()) {
-        return d_responseMessage.choice().status().code();  // RETURN
-    }
-
-    return bmqt::GenericResult::e_SUCCESS;
-}
-
-inline const bsl::string& RequestManagerRequest::nodeDescription() const
-{
-    return d_nodeDescription;
-}
-
-inline int RequestManagerRequest::groupId() const
-{
-    return d_groupId;
-}
-
-inline int RequestManagerRequest::componentId() const
-{
-    return d_componentId;
-}
-
-inline const bdld::Datum& RequestManagerRequest::userData() const
-{
-    return d_userData.datum();
-}
-
-// --------------------
-// class RequestManager
-// --------------------
-
-inline bmqt::GenericResult::Enum
-RequestManager::sendHelper(bmqio::Channel*                     channel,
-                           const bsl::shared_ptr<bdlbb::Blob>& blob_sp,
-                           bsls::Types::Int64                  watermark)
-{
-    bmqio::Status status;
-    channel->write(&status, *blob_sp, watermark);
-
-    switch (status.category()) {
-    case bmqio::StatusCategory::e_SUCCESS:
-        return bmqt::GenericResult::e_SUCCESS;
-    case bmqio::StatusCategory::e_CONNECTION:
-        return bmqt::GenericResult::e_NOT_CONNECTED;
-    case bmqio::StatusCategory::e_LIMIT:
-        return bmqt::GenericResult::e_NOT_READY;
-    case bmqio::StatusCategory::e_GENERIC_ERROR:
-    case bmqio::StatusCategory::e_TIMEOUT:
-    case bmqio::StatusCategory::e_CANCELED:
-    default: return bmqt::GenericResult::e_UNKNOWN;
-    }
-}
-
-inline void RequestManager::onRequestTimeout(int requestId)
-{
-    // executed by the thread selected by 'd_executor'
-
-    RequestSp request;
-
-    {
-        bslmt::LockGuard<bslmt::Mutex> guard(&d_mutex);  // MUTEX LOCKED
-
-        RequestMapIter it = d_requests.find(requestId);
-
-        if (it == d_requests.end()) {
-            // The request must have completed at the same time from a
-            // different thread while we were waiting on the 'd_mutex'.
-            return;  // RETURN
-        }
-
-        request = it->second;
-
-        request->d_haveTimeout = true;
-        // Do not remove the request from the map yet (a response will
-        // eventually be received, or the request be canceled).
-
-        // Explicitly invalidate the timeout since we processed it
-        request->d_timeoutSchedulerHandle.release();
-
-        if (!d_lateResponseMode) {
-            d_requests.erase(it);
-        }
-    }  // close guard scope
-
-    BALL_LOG_ERROR << "Request with '" << request->nodeDescription()
-                   << "' has timed out: " << request->request();
-
-    // Now prepare a response and invoke the callback/signal outside the mutex.
-    bmqp_ctrlmsg::ControlMessage& response = request->response();
-
-    // 1. 'fake' a response, with a Timeout status type
-    response.rId().makeValue(requestId);
-
-    bmqu::MemOutStream os;
-    os << "The request timedout after "
-       << bmqu::PrintUtil::prettyTimeInterval(
-              bmqu::Time::highResolutionTimer() - request->d_sendTime);
-
-    response.choice().makeStatus();
-    response.choice().status().code() = k_CODE_TIMEOUT_LOCAL;
-    response.choice().status().message().assign(os.str().data(),
-                                                os.str().length());
-    reinterpret_cast<int&>(response.choice().status().category()) =
-        static_cast<int>(bmqt::GenericResult::e_TIMEOUT);
-    // Note that above reinterpret & static casts are needed to suppress
-    // compiler diagnostics, because this component does not include
-    // bmqp_ctrlmsg_messages.h on purpose.
-
-    // The lateResponseMode assumes that 'onRequestTimeout' is serialized with
-    // 'processResponse' and 'cancelAllRequests' (by using 'd_executor').
-    // Meaning, the executor is responsible for ignoring timeouts after
-    // responses.
-
-    // 2. Invoke the response callback/signal: If a response callback was
-    //    provided invoke it now, otherwise signal: normally 'signal()' is
-    //    called by the caller from its response callback; however, it is
-    //    convenient to be able to write a synchronous call in a single method,
-    //    with no response callback provided, so we invoke 'signal()' ourself
-    //    now for that matter.
-
-    if (request->d_responseCb) {
-        bslma::ManagedPtr<void> spanToken(request->activateDTSpan());
-        request->d_responseCb(request);
-    }
-    else {
-        request->signal();
-    }
-}
-
-inline void
-RequestManager::applyResponse(const RequestSp&                    request,
-                              const bmqp_ctrlmsg::ControlMessage& response)
-{
-    // mutex *NOT* locked
-
-    // Cancel the timeout event
-    d_scheduler_p->cancelEvent(&(request->d_timeoutSchedulerHandle));
-
-    // Populate response field.
-
-    // The lateResponseMode assumes that 'onRequestTimeout' is serialized with
-    // 'processResponse' and 'cancelAllRequests' (by using 'd_executor').
-    // Meaning, the response field is (re)set and accessed in the same thread.
-
-    request->response() = response;
-
-    if (request->response().rId().isNull()) {
-        // The 'id' field of 'response' is not populated (can happen when we
-        // cancelAllRequests by injecting one response object) so we populate
-        // it using the 'id' of the original request.
-        BSLS_ASSERT_OPT(!request->request().rId().isNull());
-        request->response().rId().makeValue(request->request().rId().value());
-    }
-
-    // Convert timeout code: If an incoming Status response has a timeout
-    // category, then set the code to the remote timeout code
-    if (BSLS_PERFORMANCEHINT_PREDICT_UNLIKELY(
-            response.choice().isStatusValue() &&
-            (response.choice().status().category() ==
-             bmqp_ctrlmsg::StatusCategory::E_TIMEOUT))) {
-        BSLS_PERFORMANCEHINT_UNLIKELY_HINT;
-        request->response().choice().status().code() = k_CODE_TIMEOUT_REMOTE;
-    }
-
-    // Invoke the response callback/signal: If a response callback was provided
-    // invoke it now, otherwise signal: normally 'signal()' is called by the
-    // caller from its response callback; however, it is convenient to be able
-    // to write a synchronous call in a single method, with no response
-    // callback provided, so we invoke 'signal()' ourself now for that matter.
-    if (request->d_responseCb) {
-        bslma::ManagedPtr<void> spanToken(request->activateDTSpan());
-        request->d_responseCb(request);
-    }
-    else {
-        request->signal();
-    }
-}
-
-inline RequestManager::RequestManager(bmqp::EventType::Enum  eventType,
-                                      BlobSpPool*            blobSpPool_p,
-                                      bdlmt::EventScheduler* scheduler,
-                                      bool                   lateResponseMode,
-                                      bslma::Allocator*      allocator)
-: d_allocator_p(allocator)
-, d_eventType(eventType)
-, d_scheduler_p(scheduler)
-, d_nextRequestId(0)
-, d_requests(allocator)
-, d_schemaEventBuilder(blobSpPool_p, bmqp::EncodingType::e_BER, allocator)
-, d_lateResponseMode(lateResponseMode)
-, d_executor(bmqex::SystemExecutor())  // Use SystemExecutor so that when using
-                                       // 'possiblyBlocking' it will inline
-                                       // invoke the function in the callers
-                                       // thread instead of spawning a new
-                                       // thread every time.
-, d_dtContext_sp(NULL)
-{
-    // PRECONDITIONS
-    BSLS_ASSERT_SAFE(d_scheduler_p->clockType() ==
-                     bsls::SystemClockType::e_MONOTONIC);
-}
-
-inline RequestManager::RequestManager(bmqp::EventType::Enum  eventType,
-                                      BlobSpPool*            blobSpPool_p,
-                                      bdlmt::EventScheduler* scheduler,
-                                      bool                   lateResponseMode,
-                                      const bmqex::Executor& executor,
-                                      bslma::Allocator*      allocator)
-: d_allocator_p(allocator)
-, d_eventType(eventType)
-, d_scheduler_p(scheduler)
-, d_nextRequestId(0)
-, d_requests(allocator)
-, d_schemaEventBuilder(blobSpPool_p, bmqp::EncodingType::e_BER, allocator)
-, d_lateResponseMode(lateResponseMode)
-, d_executor(executor)
-, d_dtContext_sp(NULL)
-{
-    // PRECONDITIONS
-    BSLS_ASSERT_SAFE(d_scheduler_p->clockType() ==
-                     bsls::SystemClockType::e_MONOTONIC);
-    BSLS_ASSERT_SAFE(executor);
-}
-
-inline RequestManager::RequestManager(bmqp::EventType::Enum  eventType,
-                                      BlobSpPool*            blobSpPool_p,
-                                      bdlmt::EventScheduler* scheduler,
-                                      bool                   lateResponseMode,
-                                      const bmqex::Executor& executor,
-                                      const DTContextSp&     dtContext,
-                                      bslma::Allocator*      allocator)
-: d_allocator_p(allocator)
-, d_eventType(eventType)
-, d_scheduler_p(scheduler)
-, d_nextRequestId(0)
-, d_requests(allocator)
-, d_schemaEventBuilder(blobSpPool_p, bmqp::EncodingType::e_BER, allocator)
-, d_lateResponseMode(lateResponseMode)
-, d_executor(executor)
-, d_dtContext_sp(dtContext)
-{
-    // PRECONDITIONS
-    BSLS_ASSERT_SAFE(d_scheduler_p->clockType() ==
-                     bsls::SystemClockType::e_MONOTONIC);
-    BSLS_ASSERT_SAFE(executor);
-}
-
-inline RequestManager::~RequestManager()
-{
-    // PRECONDITIONS
-    BSLS_ASSERT_SAFE(d_requests.empty() &&
-                     "There are still outstanding requests, "
-                     "'cancelAllRequests()' must be called before destroying "
-                     "this object");
-}
-
-inline RequestManager&
-RequestManager::setExecutor(const bmqex::Executor& executor)
-{
-    d_executor = executor;
-    return *this;
-}
-
-inline RequestManager::RequestSp RequestManager::createRequest()
-{
-    RequestSp request;
-    request.createInplace(d_allocator_p, d_allocator_p);
-    request->d_self_wp =
-        request;  // Give request a 'weak_ptr' to itself so it
-                  // can get back a 'shared_ptr' in 'signal()'.
-    if (d_dtContext_sp) {
-        request->setDTContext(d_dtContext_sp.get());
-    }
-
-    return request;
-}
-
-inline bmqt::GenericResult::Enum
-RequestManager::sendRequest(const RequestSp&          request,
-                            bmqio::Channel*           channel,
-                            const bsl::string&        nodeDescription,
-                            const bsls::TimeInterval& timeout,
-                            bsls::Types::Int64        watermark,
-                            bsl::string*              errorDescription)
-{
-    return sendRequest(request,
-                       bdlf::BindUtil::bind(&sendHelper,
-                                            channel,
-                                            bdlf::PlaceHolders::_1,  // blob
-                                            watermark),  // watermark
-                       nodeDescription,
-                       timeout,
-                       errorDescription);
-}
-
-inline bmqt::GenericResult::Enum
-RequestManager::sendRequest(const RequestSp&          request,
-                            const SendFn&             sendFn,
-                            const bsl::string&        nodeDescription,
-                            const bsls::TimeInterval& timeout,
-                            bsl::string*              errorDescription)
-{
-    bslmt::LockGuard<bslmt::Mutex> guard(&d_mutex);  // MUTEX LOCKED
-
-    // Inject the requestId in the request
-    int requestId = ++d_nextRequestId;
-    request->request().rId().makeValue(requestId);
-
-    request->d_nodeDescription = nodeDescription;
-
-    // Prepare the message to send
-    d_schemaEventBuilder.reset();
-    int rc = d_schemaEventBuilder.setMessage(request->request(), d_eventType);
-    if (rc != 0) {
-        bmqu::MemOutStream errorDesc;
-        errorDesc << "ENCODING_FAILED, rc: " << rc;
-        if (errorDescription) {
-            *errorDescription = errorDesc.str();
-        }
-        else {
-            BALL_LOG_ERROR << "Unable to send request to '" << nodeDescription
-                           << "' [reason: " << errorDesc.str()
-                           << "]: " << request->request();
-        }
-        return bmqt::GenericResult::e_INVALID_ARGUMENT;  // RETURN
-    }
-
-    BALL_LOG_INFO << "Sending request to '" << nodeDescription << "' "
-                  << "[request: " << request->request()
-                  << ", timeout: " << timeout << "]";
-
-    // We are under the 'd_mutex' lock, so we have guarantee that even if the
-    // response comes back before we added the request to the map, it won't be
-    // processed until we return from this method, meaning we registered it to
-    // the map.
-
-    // Send the request
-    request->d_sendTime              = bmqu::Time::highResolutionTimer();
-    bmqt::GenericResult::Enum sendRc = sendFn(d_schemaEventBuilder.blob());
-    if (sendRc != bmqt::GenericResult::e_SUCCESS) {
-        bmqu::MemOutStream errorDesc;
-        errorDesc << "WRITE_FAILED, status: " << sendRc;
-        if (errorDescription) {
-            *errorDescription = errorDesc.str();
-        }
-        else {
-            BALL_LOG_ERROR << "Unable to send request to '" << nodeDescription
-                           << "' [reason: " << errorDesc.str()
-                           << "]: " << request->request();
-        }
-        return sendRc;  // RETURN
-    }
-
-    // Schedule a timeout
-    struct Local {
-        static void dispatch(bmqex::Executor*             executor,
-                             const bsl::function<void()>& callback)
-        {
-            executor->dispatch(callback);
-        }
-    };
-
-    d_scheduler_p->scheduleEvent(
-        &(request->d_timeoutSchedulerHandle),
-        bmqu::Time::nowMonotonicClock() + timeout,
-        bdlf::BindUtil::bind(&Local::dispatch,
-                             &d_executor,
-                             bsl::function<void()>(bdlf::BindUtil::bind(
-                                 &RequestManager::onRequestTimeout,
-                                 this,
-                                 requestId))));
-
-    // Insert the request in the map
-    BSLA_MAYBE_UNUSED bsl::pair<RequestMapIter, bool> insertRC =
-        d_requests.insert(bsl::make_pair(requestId, request));
-    BSLS_ASSERT_SAFE(insertRC.second);
-
-    return bmqt::GenericResult::e_SUCCESS;
-}
-
-inline int
-RequestManager::processResponse(const bmqp_ctrlmsg::ControlMessage& response)
-{
-    enum RcEnum {
-        // Value for the various RC error categories
-        rc_SUCCESS   = 0,
-        rc_NOT_FOUND = -1
-    };
-
-    RequestSp request;
-
-    {
-        bslmt::LockGuard<bslmt::Mutex> guard(&d_mutex);  // MUTEX LOCKED
-
-        RequestMapIter it = d_requests.find(response.rId().value());
-
-        if (it == d_requests.end()) {
-            // The request must have completed at the same time from a
-            // different thread while we were waiting on the 'd_mutex'.
-            BALL_LOG_DEBUG << "Received a response for a non existent request"
-                           << ", dropping it";
-            return rc_NOT_FOUND;  // RETURN
-        }
-
-        request = it->second;
-        d_requests.erase(it);
-
-        if (request->d_haveTimeout && !d_lateResponseMode) {
-            // Ignore late response
-            BALL_LOG_DEBUG << "Ignoring late response: " << response;
-            return rc_NOT_FOUND;  // RETURN
-        }
-
-        request->d_haveResponse = true;
-    }
-
-    applyResponse(request, response);
-
-    return rc_SUCCESS;
-}
-
-inline void
-RequestManager::cancelAllRequests(const bmqp_ctrlmsg::ControlMessage& reason)
-{
-    cancelAllRequestsImpl(reason,
-                          RequestType::k_NO_GROUP_ID,
-                          RequestManagerComponentId::k_NO_COMPONENT_ID);
-}
-
-inline void
-RequestManager::cancelGroupRequests(const bmqp_ctrlmsg::ControlMessage& reason,
-                                    int groupId)
-{
-    // PRECONDITIONS
-    BSLS_ASSERT_SAFE(groupId != RequestType::k_NO_GROUP_ID);
-
-    cancelAllRequestsImpl(reason,
-                          groupId,
-                          RequestManagerComponentId::k_NO_COMPONENT_ID);
-}
-
-inline void RequestManager::cancelComponentRequests(
-    const bmqp_ctrlmsg::ControlMessage& reason,
-    int                                 componentId)
-{
-    // PRECONDITIONS
-    BSLS_ASSERT_SAFE(componentId !=
-                     RequestManagerComponentId::k_NO_COMPONENT_ID);
-
-    cancelAllRequestsImpl(reason, RequestType::k_NO_GROUP_ID, componentId);
-}
-
-inline void RequestManager::cancelAllRequestsImpl(
-    const bmqp_ctrlmsg::ControlMessage& reason,
-    int                                 groupId,
-    int                                 componentId)
-{
-    // Note that requests must be cancelled in the same order in which they
-    // were sent.
-
-    // Create a new map so we can work on it outside the mutex
-    RequestMap requestsCopy(d_requests.bucket_count(),
-                            d_requests.get_allocator());
-
-    {
-        bslmt::LockGuard<bslmt::Mutex> guard(&d_mutex);  // MUTEX LOCKED
-
-        RequestMapIter it = d_requests.begin();
-        while (it != d_requests.end()) {
-            const bool matchGroup = (groupId == RequestType::k_NO_GROUP_ID) ||
-                                    (groupId == it->second->groupId());
-            const bool matchComponent =
-                (componentId ==
-                 RequestManagerComponentId::k_NO_COMPONENT_ID) ||
-                (componentId == it->second->componentId());
-
-            if (matchGroup && matchComponent) {
-                // Do not notify about timed out requests.
-                if (!it->second->d_haveTimeout) {
-                    BSLA_MAYBE_UNUSED bsl::pair<RequestMapIter, bool>
-                                      insertRC = requestsCopy.insert(
-                            bsl::make_pair(it->first, it->second));
-                    BSLS_ASSERT_SAFE(insertRC.second);
-                }
-                d_requests.erase(it++);
-            }
-            else {
-                ++it;
-            }
-        }
-    }
-
-    const bool hasGroup     = (groupId != RequestType::k_NO_GROUP_ID);
-    const bool hasComponent = (componentId !=
-                               RequestManagerComponentId::k_NO_COMPONENT_ID);
-    if (!hasGroup && !hasComponent) {
-        BALL_LOG_INFO << "Canceling all requests (" << requestsCopy.size()
-                      << " items) with " << reason << ".";
-    }
-    else if (hasGroup && !hasComponent) {
-        BALL_LOG_INFO << "Canceling requests belonging to group '" << groupId
-                      << "' (" << requestsCopy.size() << " items) with "
-                      << reason << ".";
-    }
-    else if (!hasGroup && hasComponent) {
-        BALL_LOG_INFO << "Canceling requests belonging to component '"
-                      << componentId << "' (" << requestsCopy.size()
-                      << " items) with " << reason << ".";
-    }
-    else {
-        BALL_LOG_INFO
-            << "Canceling requests simultaneously belonging to group '"
-            << groupId << "' and component '" << componentId << "' ("
-            << requestsCopy.size() << " items) with " << reason << ".";
-    }
-
-    if (requestsCopy.empty()) {
-        return;  // RETURN
-    }
-
-    for (RequestMapConstIter it = requestsCopy.begin();
-         it != requestsCopy.end();
-         ++it) {
-        applyResponse(it->second, reason);
-    }
-}
 
 }  // close package namespace
 }  // close enterprise namespace
