+++ src/groups/mqb/mqba/mqba_clientsession.cpp
@@ -185,6 +185,7 @@
 #include <bmqtsk_alarmlog.h>
 #include <bmqu_blob.h>
 #include <bmqu_printutil.h>
+#include <bmqu_time.h>
 #include <bmqu_weakmemfn.h>
 
 // BDE
