+++ src/groups/mqb/mqbnet/mqbnet_multirequestmanager.t.cpp
@@ -30,6 +30,7 @@
 #include <mqbnet_mockcluster.h>
 
 #include <bmqu_tempdirectory.h>
+#include <bmqu_time.h>
 
 // BDE
 #include <bdlbb_blobutil.h>
